+++ defaults/public/js/overlook/panelCollapse.js
@@ -1,0 +1,17 @@
+/* global $ */
+/* global document */
+
+// From: http://bootsnipp.com/snippets/featured/collapsible-panel
+
+$(document).on('click', '.panel-collapse-button', function() {
+    var $this = $(this);
+	if(!$this.hasClass('panel-collapsed')) {
+		$this.parents('.panel').find('.panel-body').slideUp();
+		$this.addClass('panel-collapsed');
+		$this.find('i').removeClass('glyphicon-chevron-up').addClass('glyphicon-chevron-down');
+	} else {
+		$this.parents('.panel').find('.panel-body').slideDown();
+		$this.removeClass('panel-collapsed');
+		$this.find('i').removeClass('glyphicon-chevron-down').addClass('glyphicon-chevron-up');
+	}
+});